+++ src/sisl/io/siesta/_help.py
@@ -7,6 +7,7 @@ import sisl._array as _a
 from sisl import SislError
 from sisl._core.sparse import _rows_and_cols
 from sisl.messages import warn
+from sisl.sparse import _to_coo
 
 try:
     from . import _siesta